Create ShrinkText compopnent
Created a ShrinkText component. Also Check proptypes in ShrinkText and TextInput components.
